+++ biopipen/scripts/tcr/CDR3AAPhyschem.R
@@ -170,6 +170,8 @@ immdata = readRDS(immdatafile)
 merge_data = function(sam) {
     # Merge the data for one sample from immdata and metadata
     out = immdata$data[[sam]] %>%
+        separate_rows(chain, CDR3.aa, V.name, J.name, sep = ";") %>%
+        filter(chain == "TRB") %>%
         mutate(
             Sample = sam,
             locus = "TCRB",